animal-face-id/requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing & scientific computing
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Configuration & serialization
pyyaml>=6.0
joblib>=1.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0

# UI
gradio>=4.0.0

